+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+* Raise a ValidationError on Movement23 if the counterparty
+  currency is not valid
+
 ## 1.0.1 - 2021-11-25
 
 * Release for testing github actions